32-bit support

+++ src/termios.rs
@@ -1,10 +1,10 @@
 use libc::{c_int, c_uint, c_uchar};
 
 #[cfg(not(target_os = "macos"))]
-pub const TIOCGWINSZ: u64 = 0x00005413;
+pub const TIOCGWINSZ: usize = 0x00005413;
 
 #[cfg(target_os = "macos")]
-pub const TIOCGWINSZ: u64 = 0x40087468;
+pub const TIOCGWINSZ: usize = 0x40087468;
 
 extern {
     pub fn tcgetattr(filedes: c_int, termptr: *mut Termios) -> c_int;
